Add support for ctxLink. bmdebug: handle "help" command for find and trace. decodectf: BUG fix in array overrun (FIFO queue). parsetsdl: check for missing stream definition (missing event header). swotrace: BUG fix in setting timestamps when there are multiple CTF trace packets in a single USB frame.

+++ source/nuklear_gdip.c
@@ -590,8 +590,8 @@ nk_gdip_image_to_nk(GpImage *image) {
     img = nk_image_ptr( (void*)image );
     GdipGetImageHeight(image, &uheight);
     GdipGetImageWidth(image, &uwidth);
-    img.h = uheight;
-    img.w = uwidth;
+    img.h = (unsigned short)uheight;
+    img.w = (unsigned short)uwidth;
     return img;
 }
